Revert "Avoid writing config file if not installed (#26107) (#26113)"
This reverts commit 78722734fe96ec660993191e622459a600482004.

It does not create `LFS_JWT_SECRET` if `INSTALL_LOCK` is true and the
value of `LFS_JWT_SECRET` found in `app.ini` is incorrect. As a result
LFS_JWT_SECRET will not be set at all and the Forgejo admin will not
be notified that the value in the `app.ini` was ignored.